+++ spec/unit/test_reader.rb
@@ -5,7 +5,7 @@ module Alf
     subject{ Reader }
     it { should respond_to(:rash) }
       
-    describe "rash" do
+    describe "rash on a real IO" do
       subject{ Reader.rash($stdin) }
       it{ should be_a(Reader::Rash) }
     end
@@ -23,6 +23,10 @@ module Alf
       
       specify "when an IO" do
         Reader.reader($stdin).should be_a(Reader::Rash)
+      end
+      
+      specify "when an StringIO" do
+        Reader.reader(StringIO.new).should be_a(Reader::Rash)
       end
       
       specify "with options" do